requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.6
pydantic==2.10.3
pydantic-settings==2.7.0

# Document Processing
numpy>=1.26.0,<2.0.0
pdfplumber==0.10.3
PyPDF2==3.0.1
opencv-python-headless==4.9.0.80
pdf2image==1.17.0
Pillow>=10.3.0
pytesseract==0.3.10
reportlab==4.0.9
matplotlib>=3.7.0

# AI/ML
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
